scheduler_backend/pyproject.toml
[tool.poetry]
name = "scheduler-backend"
version = "0.1.0"
description = "backend service for TG_scheduler"
authors = ["mssffct <herooftheday82@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
djoser = "^2.2.2"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
django-cleanup = "^8.0.0"
redis = "^5.0.1"
channels-redis = "^4.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
whitenoise = "^6.6.0"
django-rest-framework = "^0.1.0"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
psycopg2 = "^2.9.9"
django-filter = "^23.4"
daphne = "^4.0.0"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
